requirements.txt
anymarkup
redis
celery[eventlet,gevent]
jinja2
jsonschema
pyyaml
python-qpid-proton
raven
requests
